backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
psycopg2-binary>=2.9.1
pokerkit
python-dotenv>=0.19.0
pydantic>=1.8.2
pydantic-settings